# marketplace-api-contract/src/main/resources/marketplace-api-contract.yaml
openapi: 3.0.3
info:
  title: OnlyDust Marketplace API
  description: OnlyDust Marketplace API
  version: 1.0.0
servers:
  - url: 'http://localhost:9999'
    description: Local
  - url: 'https://develop-api.onlydust.com'
    description: Develop
  - url: 'https://staging-api.onlydust.com'
    description: Staging
  - url: 'https://api.onlydust.com'
    description: Production
paths:
...
  /api/v1/projects:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Create project
      operationId: createProject
      description: |
        Create new project. The caller must be a leader of the project.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "200":
          description: 'Create Project response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProjectResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    get:
      security:
        - { }
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get all projects
      operationId: getProjects
      description: |
        Get the (soon to be paginated) list of all projects on the platform that are visible to the caller.
      parameters:
        - in: query
          name: sort
          description: Sort order
          schema:
            type: string
            enum:
              - RANK
              - NAME
              - REPO_COUNT
              - CONTRIBUTOR_COUNT
        - in: query
          name: technologies
          description: Technology filter
          schema:
            type: array
            items:
              type: string
        - in: query
          name: sponsor
          description: Sponsor filter
          schema:
            type: array
            items:
              type: string
        - in: query
          name: mine
          description: Filter mine projects
          schema:
            type: boolean
        - in: query
          name: search
          description: Full text search
          schema:
            type: string
      responses:
        "200":
          description: 'List of projects'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}:
    get:
      security:
        - { }
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project details
      operationId: getProject
      description: |
        Get the project if it is visible to the caller.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: query
          name: includeAllAvailableRepos
          schema:
            type: boolean
      responses:
        "200":
          description: 'The project'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Update project
      operationId: updateProject
      description: |
        Update project. The caller must be a leader of the project.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        "204":
          description: 'The project was updated successfully.'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/slug/{slug}:
    get:
      security:
        - { }
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project details
      operationId: getProjectBySlug
      description: |
        Get the project if it is visible to the caller.
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
        - in: query
          name: includeAllAvailableRepos
          schema:
            type: boolean
      responses:
        "200":
          description: 'The project'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/contributors:
    get:
      security:
        - { }
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get all project contributors
      operationId: getProjectContributors
      description: |
        Get the (soon to be paginated) list of all contributors of the project.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: query
          name: sort
          description: Sort order
          schema:
            type: string
            enum:
              - LOGIN
              - CONTRIBUTION_COUNT
              - REWARD_COUNT
              - EARNED
              - TO_REWARD_COUNT
        - in: query
          name: direction
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - in: query
          name: pageIndex
          required: true
          schema:
            type: integer
        - in: query
          name: pageSize
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 'List of project contributors'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributorsPageResponse'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/search/contributors:
    get:
      security:
        - { }
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Search project contributors by login
      operationId: searchProjectContributors
      description: |
        Get the list of all contributors of the project matching the input query.
        If not enough results are found, the search is extended to the 5 best matching github users.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: query
          name: login
          description: login query
          schema:
            type: string
      responses:
        "200":
          description: 'List of project contributors'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributorSearchResponse'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/budgets:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project budgets
      operationId: getProjectBudgets
      description: |
        Get the list of all budgets from this project. Only project leaders can access this information.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
      responses:
        "200":
          description: 'List of budgets'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectBudgetsResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/rewards:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project rewards
      operationId: getProjectRewards
      description: |
        Get the list of all rewards granted from this project. Only project leaders can access this information.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: query
          name: sort
          description: Sort order
          schema:
            enum:
              - REQUESTED_AT
              - CONTRIBUTION
              - AMOUNT
              - STATUS
        - in: query
          name: direction
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - in: query
          name: pageIndex
          required: true
          schema:
            type: integer
        - in: query
          name: pageSize
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 'List of rewards'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardsPageResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Create reward
      operationId: createReward
      description: |
        Create new reward. The caller must be a leader of the project.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
      requestBody:
        description: 'New reward'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardRequest'
      responses:
        "200":
          description: 'Reward'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRewardResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/rewards/{rewardId}:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project reward by reward Id
      operationId: getProjectReward
      description: |
        Get reward details from this project. Only project leaders can access this information.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: path
          name: rewardId
          required: true
          schema:
            $ref: '#/components/schemas/RewardId'
      responses:
        "200":
          description: 'Reward details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/RewardForbiddenError'
        "404":
          $ref: '#/components/responses/RewardNotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'

    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Create reward
      operationId: cancelReward
      description: |
        Create new reward. The caller must be a leader of the project.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: path
          name: rewardId
          required: true
          schema:
            $ref: '#/components/schemas/RewardId'
      responses:
        "204":
          description: 'Reward deleted'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/ProjectForbiddenError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/projects/{projectId}/rewards/{rewardId}/reward-items:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - Projects
      summary: Get project reward items page by reward Id
      operationId: getProjectRewardItemsPage
      description: |
        Get reward items details from this project. Only project leaders can access this information.
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            $ref: '#/components/schemas/ProjectId'
        - in: path
          name: rewardId
          required: true
          schema:
            $ref: '#/components/schemas/RewardId'
        - in: query
          name: pageIndex
          description: Page number
          schema:
            type: integer
        - in: query
          name: pageSize
          description: Number of items returned by page
          schema:
            type: integer
      responses:
        "200":
          description: 'Reward details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardItemsPageResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/RewardForbiddenError'
        "404":
          $ref: '#/components/responses/RewardNotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/projects/logos:
    post:
      security:
        - { }
      tags:
        - Projects
      summary: Upload a project logo
      operationId: uploadProjectLogo
      requestBody:
        content:
          image/png:
            schema:
              type: string
              format: binary
          image/jpeg:
            schema:
              type: string
              format: binary
          image/svg+xml:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: 'Image uploaded'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadImageResponse'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          example:
            message: Github API is not responding
            status: 500
            id: 4f7c660e-6cf1-4424-b240-e630c4c19069
          schema:
            $ref: '#/components/schemas/OnlyDustError'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          example:
            message: You must be authenticated to make this call
            status: 401
            id: 4f7c660e-6cf1-4424-b240-e630c4c19069
          schema:
            $ref: '#/components/schemas/OnlyDustError'
    NotFoundError:
      description: Requested resource not found
      content:
        application/json:
          example:
            message: NOT_FOUND
            status: 404
            id: 4f7c660e-6cf1-4424-b240-e630c4c19069
          schema:
            $ref: '#/components/schemas/OnlyDustError'
    ProjectForbiddenError:
      description: Forbidden. The project is not visible to the caller.
      content:
        application/json:
          example:
            message: Access to this project is forbidden
            status: 403
            id: 4f7c660e-6cf1-4424-b240-e630c4c19069
          schema:
            $ref: '#/components/schemas/OnlyDustError'
...
  schemas:
...
    UserId:
      type: string
      format: uuid
      description: OnlyDust user ID
...
    ProjectId:
      type: string
      format: uuid
      description: OnlyDust project ID
...
    ContributorSearchResponse:
      type: object
      required:
        - internalContributors
        - externalContributors
      properties:
        internalContributors:
          type: array
          items:
            $ref: '#/components/schemas/ContributorSearchItemResponse'
        externalContributors:
          type: array
          items:
            $ref: '#/components/schemas/ContributorSearchItemResponse'
...
    ProjectListResponse:
      type: object
      required:
        - projects
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectListItemResponse'
        technologies:
          type: array
          items:
            type: string
        sponsors:
          type: array
          items:
            type: string
...
    ProjectResponse:
      type: object
      required:
        - id
        - slug
        - name
        - shortDescription
        - longDescription
        - logoUrl
        - moreInfoUrl
        - hiring
        - visibility
        - contributorCount
        - topContributors
        - leaders
        - invitedLeaders
        - sponsors
        - technologies
        - remainingUsdBudget
      properties:
        id:
          $ref: '#/components/schemas/ProjectId'
        slug:
          $ref: '#/components/schemas/ProjectSlug'
        name:
          type: string
          example: Verkle Tries
        shortDescription:
          type: string
          example: A short project description
        longDescription:
          type: string
          example: A lonnnnnnnnnnnnng project description
        logoUrl:
          type: string
          format: url
          description: Direct URL to project's logo image
          example: 'https://onlydust-app-images.s3.eu-west-1.amazonaws.com/2529199823275297272.jpg'
        moreInfoUrl:
          type: string
          format: url
          description: External link to get more information about the project
          example: 'https://t.me/my-super-project'
        hiring:
          type: boolean
          description: Is the project currently hiring?
        visibility:
          $ref: '#/components/schemas/ProjectVisibility'
        contributorCount:
          type: integer
          description: Total number of contributors within the project
          example: 163
        topContributors:
          type: array
          description: The top 3 contributors (by contribution count) of this project
          maxItems: 3
          items:
            $ref: '#/components/schemas/UserLinkResponse'
        # For now, we let the flatten repos list here to avoid breaking changes. However, it is marked as deprecated.
        # The 'organizations' field should be used instead.
        repos:
          deprecated: true
          type: array
          items:
            $ref: '#/components/schemas/GithubRepoResponse'
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/ProjectGithubOrganizationResponse'
        leaders:
          type: array
          items:
            $ref: '#/components/schemas/RegisteredUserLinkResponse'
        invitedLeaders:
          type: array
          items:
            $ref: '#/components/schemas/RegisteredUserLinkResponse'
        sponsors:
          type: array
          items:
            $ref: '#/components/schemas/SponsorResponse'
        technologies:
          $ref: '#/components/schemas/Technologies'
        remainingUsdBudget:
          type: number
          example: 20.0
        rewardSettings:
          $ref: '#/components/schemas/ProjectRewardSettings'
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        shortDescription:
          type: string
        longDescription:
          type: string
        moreInfo:
          type: array
          items:
            $ref: '#/components/schemas/MoreInfo'
        isLookingForContributors:
          type: boolean
        inviteGithubUserIdsAsProjectLeads:
          type: array
          items:
            $ref: '#/components/schemas/GithubUserId'
        githubRepoIds:
          type: array
          items:
            $ref: '#/components/schemas/GithubRepoId'
        logoUrl:
          type: string
          format: url
          description: Direct URL to project's logo image
    UpdateProjectRequest:
      type: object
      required:
        - name
        - shortDescription
        - longDescription
        - moreInfo
        - isLookingForContributors
        - logoUrl
      properties:
        name:
          type: string
        shortDescription:
          type: string
        longDescription:
          type: string
        moreInfo:
          type: array
          items:
            $ref: '#/components/schemas/MoreInfo'
        isLookingForContributors:
          type: boolean
        inviteGithubUserIdsAsProjectLeads:
          type: array
          description: List of Github user ids to invite as project leads.
            All other project lead invitations will be removed.
            If empty, all project lead invitations will be removed.
            If null, no change will be made.
            It is allowed to add a new project lead invitation directly through this list.
          items:
            $ref: '#/components/schemas/GithubUserId'
        projectLeadsToKeep:
          type: array
          description: List of project leads to keep.
            All other project leads will be removed.
            If empty, all project leads will be removed.
            If null, no change will be made.
            It is not allowed to add a new project lead directly through this list, as they must be invited first.
          items:
            $ref: '#/components/schemas/UserId'
        logoUrl:
          type: string
          format: url
          description: Direct URL to project's logo image
        rewardSettings:
          $ref: '#/components/schemas/ProjectRewardSettings'
        githubRepoIds:
          type: array
          description: List of Github repos to include in the project.
            All other repos will be unlinked from the project.
            If empty, all repos will be unlinked from the project.
            If null, no change will be made.
            It is allowed to link a new repo directly through this list.
          items:
            $ref: '#/components/schemas/GithubRepoId'
...
    CreateProjectResponse:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
...